Do not declare exist/shutdown without ProviderID
This change ensures that CCM does not declar nodes as running/shutdown
or existing/missing, if there is no ProviderID found on the node.

Queried with such nodes, the CCM will now return an error.

This ensures that the node lifecycle controller, that calls these
functions, does not act unless a ProviderID is set.

This prevents a badly named server/node, that does not
have a ProviderID, from falsly being declared as shutdown or missing.

In practice, nodes get a ProviderID from the CCM as soon as they join,
so this should have no real effect, but may act as a failsafe when
other things go wrong.