Avoid unnecesary LB creation if UUID is missing
When the k8s.cloudscale.ch/loadbalancer-uuid annotation was present on
a service, but the UUID could not be found in via API, the CCM would
start creating LBs in a loop until it was killed.

The cause of this was the LBs were looked up. The CCM aassumed that if
a UUID could not be found, that the LB must not exist, and tried to
create one. This would happen in a loop without any way for the CCM to
detect that it should stop.

This assumption has now been changed. We now check for the LB both by
name and UUID, and expect to find at most one. If the service name and
the UUID both match different LBs (as might happen if one manipulates
the service annotations manually), the CCM will return an error.

Ultimately, the CCM has no state it can call its own and infers
state from the APIs it talks to (i.e., Kubernetes and the cloudscale.ch
API).

So if someone were to change both the UUID and the name of the LB using
annotations, the CCM would have no choice but to create a new LB.

Manipulating the UUID was at one point thought to be a possible feature:

One could bring an existing LB into the managament of the CCM. Though
there is no code that would prevent this, the docs no longer mention
this option, as this has the potential to cause havoc.

Changing the name is already discouraged. It is probably savest to
not touch the name, nor the UUID, during the lifetime of a service.